requirements.txt
absl-py==1.4.0
ale-py==0.8.1
alembic==1.11.1
appnope==0.1.3
asttokens==2.2.1
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
backcall==0.2.0
cachetools==5.3.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmaes==0.9.1
colorlog==6.7.0
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
Farama-Notifications==0.0.4
filelock==3.12.2
fonttools==4.39.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.54.2
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.28.1
gymnasium-notices==0.0.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
ipywidgets==8.0.4
jax-jumpy==1.0.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
optuna==3.2.0
packaging==23.0
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.4.0
pyglet==2.0.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
scipy==1.10.1
Shimmy==0.2.1
six==1.16.0
SQLAlchemy==2.0.16
stable-baselines3==2.0.0a13
stack-data==0.6.2
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
torch==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.6
widgetsnbextension==4.0.5
zipp==3.15.0
